make preconditions 'generic'
