--- src/examples/zkapps/voting/election_preconditions.ts
@@ -1,12 +1,0 @@
-import { CircuitValue, prop, UInt32 } from 'snarkyjs';
-
-export class ElectionPreconditions extends CircuitValue {
-  @prop startElection: UInt32;
-  @prop endElection: UInt32;
-
-  constructor(startElection: UInt32, endElection: UInt32) {
-    super();
-    this.startElection = startElection;
-    this.endElection = endElection;
-  }
-}
